requirements.txt
gradio
requests
python-dotenv
fpdf2